feat(FriendController, FavoriteController): guards

+++ app/Http/Controllers/FavoriteController.php
@@ -8,17 +8,25 @@ use Illuminate\Support\Facades\Auth;
 class FavoriteController extends Controller {
 	public function store(string $id) {
 		$user = Auth::user();
-		$user->favoriteWorks()->attach($id);
-
 		$work = Work::find($id);
+
+		if ($work === null) {
+			return back();
+		}
+
+		$user->favoriteWorks()->attach($id);
 		return back()->with('success', 'You have favorited "' . $work->title . '" by ' . $user->name . '.');
 	}
 
 	public function destroy(string $id) {
 		$user = Auth::user();
-		$user->favoriteWorks()->detach($id);
-
 		$work = Work::find($id);
+
+		if ($work === null) {
+			return back();
+		}
+
+		$user->favoriteWorks()->detach($id);
 		return back()->with('success', 'You have unfavorited "' . $work->title . '" by ' . $user->name . '.');
 	}
 }
